requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.31
protobuf==3.20.1
requests==2.31.0
urllib3==2.0.5
Werkzeug==3.0.0
